Refactor cache management and enhance local testing of cache backends
- Moved `Caches` struct to a separate module (`caches_extension.rs`) for better organization and maintainability.
- Updated `local_server_config.yaml` to default the cache backend to 'none', facilitating safer and more predictable local development testing.
- Enhanced traceability in `extractor_router.rs` with additional debug logs for cache hits and misses.
- Refactored cache initialization in `server.rs` to use the newly modularized `Caches` structure.
- Made minor adjustments to `redis_cache.rs` and `traits.rs` for improved code clarity.
- Expanded documentation in `server_config.rs` to provide clearer guidance on cache configuration options.
- These modifications are driven by the need to thoroughly test each caching backend locally, ensuring robust and reliable cache behavior in different scenarios.
