+++ local_server_config.yaml
@@ -59,9 +59,12 @@ peers:
   - node_id: 0
     addr: localhost:8950
 
+# Cache configuration. It is used for caching the results
+# of queries to the vector index.
 cache:
   # Possible values - redis, memory, none
-  backend: memory
+  # Disabled by default to prevent unexpected behavior in local development
+  backend: none
 
   redis:
     addr: redis://localhost:6379
